share MarkdownTest::check_panic
